LayoutTests/webaudio/Panner/panner-cone-gain-nan-expected.txt

Tests that precision issues do not cause ConeEffect::gain() to return NaN

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS data.some(isNaN) is false
PASS successfullyParsed is true

TEST COMPLETE
